refactor(webapp): group usage chart files and tidy BreakdownChart (EXT-1144)
Address PR review: move the chart pieces into patterns/chart/ (with an
index re-exporting ChartCard / ChartSeries / colors), de-nest the series
ternaries into early-return branches, pull the date formatters to module
level, and drop the animations note. Add a comment pointing at NAN-5942
(chart test coverage).
