+++ packages/webapp/src/components/patterns/chart/index.ts
@@ -1,0 +1,4 @@
+export { ChartCard } from './ChartCard';
+export { REST_SERIES_COLOR, REST_SERIES_KEY, colorForValue } from './usageChartColors';
+
+export type { ChartSeries } from './types';
